cyber-god/backend/requirements.txt
fastapi[standard]>=0.115.0,<1.0
pydantic>=2.7.0,<3.0
uvicorn[standard]>=0.30.0
openai~=1.102
mcp>=1.0
python-dotenv>=1.0.0
